requirements.txt
streamlit
streamlit-chat
python-dotenv
PyPDF2
python-docx
langchain
langchain-community
transformers
torch
sentence-transformers